Create README.md
About
THE FLINTSTONES TOKEN (YABBA) IS A TRIBUTE TO THE MOST POPULAR CARTOON 
OF THE LAST 60 YEARS WITH MILLIONS OF FOLLOWERS, ENTHUSIASTS AND COLLECTORS 
AROUND THE WORLD
YABBA is  founded  on  the  vision that crypto can empower everyone to reach new heights of financial freedom.
If traditional finance has left you dead, get aboard and be a part of our journey. Everybody is welcome, everybody is equal. Take you chance an prepare for the greatest adventure of our lifetimes!

https://theflintstonestoken.com/

https://bscscan.com/address/0x26386ae2c8209ac1598b2f9675909cd8bac66542